langchain-runnables-Part1/requirements.txt
# LangChain Core
langchain
langchain-core

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Environment Variable Management
python-dotenv

